requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cloudinary==1.28.0
Django==4.0
django-crispy-forms==1.13.0
Pillow==8.4.0
psycopg2-binary==2.9.2
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
